backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
databases
aiosqlite